Take out Hero and button as a reusable component  (#98)
* Take out Hero and button as a reusable component

* fixed pixel perfect things

* fixed netlify build with removing all unused variables
